updated combat function and added preliminary bard info function
there is a lot of info that is class-specific and i intend to have it all in this program at some point. but that'll be hard. maybe not worth it. for now, I added a preliminary "function" that just prints some tables with mods for listening for noises and climbing.